feat: idf-5 ble, proto string encode separation
