+++ examples/shadow_mqtt/sdkconfig
@@ -1709,17 +1709,20 @@ CONFIG_ESP_WIFI_AUTH_WPA2_PSK=y
 # CONFIG_ESP_WIFI_AUTH_WPA3_PSK is not set
 # CONFIG_ESP_WIFI_AUTH_WPA2_WPA3_PSK is not set
 # CONFIG_ESP_WIFI_AUTH_WAPI_PSK is not set
-# CONFIG_GOLAIN_BLE is not set
+CONFIG_GOLAIN_BLE=y
 # CONFIG_GOLAIN_MQTT_OTA is not set
 
 #
 # Persistent Logging
 #
-# CONFIG_GOLAIN_CLOUD_LOGGING is not set
+CONFIG_GOLAIN_CLOUD_LOGGING=y
+CONFIG_GOLAIN_MAX_PERSISTENT_LOGS=50
+CONFIG_PERSISTENT_LOGS_INTERNAL_LOG_LEVEL=0
 
 #
 # Device Health
 #
+# CONFIG_GOLAIN_REPORT_DEVICE_HEALTH is not set
 # end of Device Health
 # end of Persistent Logging
