+++ src/golain.c
@@ -23,6 +23,7 @@
 #include "golain_types.h"
 #include "golain.h"
 #include "golain_hal.h"
+#include "golain_encode.h"
 
 // #include "esp_log.h"
 
@@ -326,26 +327,4 @@ golain_err_t golain_device_health_decode_message(uint8_t *buffer, size_t message
     return status;
 }
 
-bool golain_pb_decode_string(pb_istream_t *stream, const pb_field_t *field, void **arg){
-    uint8_t buffer[1024] = {0};
-    
-    /* We could read block-by-block to avoid the large buffer... */
-    if (stream->bytes_left > sizeof(buffer) - 1)
-        return false;
-    
-    if (!pb_read(stream, buffer, stream->bytes_left))
-        return false;
-    
-    sprintf((char*)*arg, "%s", buffer);
-    return true;
-    
-}
-
-bool golain_pb_encode_string(pb_ostream_t *stream, const pb_field_t *field, void * const *arg){
-    char* string_to_encode = (char*)*arg;
-    if (!pb_encode_tag_for_field(stream, field))
-        return false;
-    
-    return pb_encode_string(stream, (uint8_t*)string_to_encode, strlen(string_to_encode));
-}
 #endif
