+++ src/golain_hal.c
@@ -23,6 +23,7 @@
 #include "golain_types.h"
 #include "golain.h"
 #include "golain_hal.h"
+#include "golain_encode.h"
 
 #include <sys/param.h>
 #include <string.h>
@@ -705,6 +706,8 @@ golain_err_t golain_hal_ble_init(golain_t* golain){
     return GOLAIN_OK;
 }
 
+#endif // CONFIG_GOLAIN_BLE
+
 /*------------------------------------------------------------------------Persistent Logs----------------------------------------------------*/
 #ifdef CONFIG_GOLAIN_CLOUD_LOGGING
 golain_err_t _golain_hal_p_log_check_nvs_errors(esp_err_t err)
@@ -887,7 +890,7 @@ golain_err_t _golain_hal_p_log_get_number_of_logs(int32_t *num)
 
     return nvs_get_i32(p_log_handle, "last_log_id", num);
 }
-#endif
+#endif // CONFIG_GOLAIN_PERSISTENT_LOGS
 /*-----------------------------------------------------------------------Device Health------------------------------------------------------*/
 #ifdef GOLAIN_REPORT_DEVICE_HEALTH
 
@@ -937,6 +940,4 @@ int8_t _golain_hal_reset_counter(void){
     return num;
     
 }
-#endif
-
-#endif
+#endif // GOLAIN_REPORT_DEVICE_HEALTH
